Moved cloneAttributes to SteppableBasePy. Made mitosis steppable base classes initialize self.parentCell and self.childCell automatically
