+++ CompuCell3D/core/pythonSetupScripts/PySteppables.py
@@ -1065,6 +1065,97 @@ class SteppableBasePy(SteppablePy,SBMLSolverHelper):
             
         return tmpElement,None                
             
+    def cloneAttributes(self,sourceCell, targetCell, no_clone_key_dict_list = [] ):
+        # clone "C++" attributes
+        from copy import deepcopy
+        for attrName in self.clonableAttributeNames:
+            setattr(targetCell , attrName, getattr(sourceCell,attrName) )
+            
+        # clone dictionary
+        for key, val in sourceCell.dict.iteritems():
+            
+            if key in no_clone_key_dict_list:
+                continue
+                
+            if key == 'SBMLSolver':
+                self.copySBMLs(_fromCell = sourceCell,_toCell = targetCell)
+            
+            if key == 'Bionetwork':
+                import bionetAPI
+                bionetAPI.copyBionetworkFromParent( sourceCell, targetCell )
+                
+            # copying the rest of dictionary entries    
+            targetCell.dict [key] = deepcopy(sourceCell.dict[key])
+            
+        # now copy data associated with plugins
+        # AdhesionFlex
+        if self.adhesionFlexPlugin:
+            sourceAdhesionVector = self.adhesionFlexPlugin.getAdhesionMoleculeDensityVector( sourceCell )
+            self.adhesionFlexPlugin.assignNewAdhesionMoleculeDensityVector( targetCell , sourceAdhesionVector )
+            
+        # PolarizationVector
+        if self.polarizationVectorPlugin:
+            sourcePolarizationVector = self.polarizationVectorPlugin.getPolarizationVector(sourceCell)
+            self.polarizationVectorPlugin.setPolarizationVector(targetCell, sourcePolarizationVector[0] , sourcePolarizationVector[1], sourcePolarizationVector[2] )
+            
+        #polarization23Plugin     
+        if self.polarization23Plugin:
+            polVec = self.polarization23Plugin.getPolarizationVector(sourceCell)
+            self.polarization23Plugin.setPolarizationVector(targetCell,polVec)
+            polMark = self.polarization23Plugin.getPolarizationMarkers(sourceCell)
+            self.polarization23Plugin.setPolarizationMarkers(targetCell , polMark[0], polMark[1] )
+            l = self.polarization23Plugin.getLambdaPolarization(sourceCell)
+            self.polarization23Plugin.setLambdaPolarization(targetCell,l)
+            
+        #CellOrientationPlugin 
+        if self.cellOrientationPlugin:
+            l = self.cellOrientationPlugin.getLambdaCellOrientation(sourceCell)
+            self.cellOrientationPlugin.setLambdaCellOrientation(targetCell,l)
+            
+        #ContactOrientationPlugin 
+        if self.contactOrientationPlugin:            
+            oVec = self.contactOrientationPlugin.getOriantationVector(sourceCell)
+            self.contactOrientationPlugin.setOriantationVector(targetCell, oVec.x, oVec.y, oVec.z)
+            self.contactOrientationPlugin.setAlpha(targetCell, self.contactOrientationPlugin.getAlpha(sourceCell) )
+            
+        #ContactLocalProductPlugin 
+        if self.contactLocalProductPlugin:
+            cVec = self.contactLocalProductPlugin.getCadherinConcentrationVec(sourceCell)
+            self.contactLocalProductPlugin.setCadherinConcentrationVec(targetCell , cVec)
+
+        #LengthConstraintPlugin    
+        if self.lengthConstraintPlugin:
+            l = self.lengthConstraintPlugin.getLambdaLength(sourceCell)        
+            tl = self.lengthConstraintPlugin.getTargetLength(sourceCell)        
+            mtl = self.lengthConstraintPlugin.getMinorTargetLength(sourceCell)    
+            self.lengthConstraintPlugin.setLengthConstraintData(targetCell, l, tl, mtl)
+            
+        #ConnectivityGlobalPlugin    
+        if self.connectivityGlobalPlugin:            
+            cs = self.connectivityGlobalPlugin.getConnectivityStrength(sourceCell)
+            self.connectivityGlobalPlugin.setConnectivityStrength(targetCell , cs)
+
+        #ConnectivityLocalFlexPlugin    
+        if self.connectivityLocalFlexPlugin:
+            cs = self.connectivityLocalFlexPlugin.getConnectivityStrength( sourceCell )
+            self.connectivityLocalFlexPlugin.setConnectivityStrength( targetCell, cs )
+            
+        #Chemotaxis    
+        if self.chemotaxisPlugin:
+            fieldNames  = self.chemotaxisPlugin.getFieldNamesWithChemotaxisData( sourceCell )
+            
+            for fieldName in fieldNames:                
+                source_chd=chemotaxisPlugin.getChemotaxisData(sourceCell,fieldName)
+                target_chd=chemotaxisPlugin.addChemotaxisData(targetCell,fieldName)
+                
+                target_chd.setLambda( source_chd.getLambda() )
+                target_chd.saturationCoef = source_chd.saturationCoef
+                target_chd.setChemotaxisFormulaByName(source_chd.formulaName)
+                target_chd.assignChemotactTowardsVectorTypes(source_chd.getChemotactTowardsVectorTypes())
+                
+        #FocalPointPLasticityPlugin - this plugin has to be handled manually - there is no good way to figure out which links shuold be copied from parent to daughter cell    
+                
+            
 
 class RunBeforeMCSSteppableBasePy(SteppableBasePy):
     def __init__(self,_simulator,_frequency=1):
